requirements.txt
pandas>=2.2.3
requests==2.31.0
bs4==0.0.2
vnai>=2.0
openpyxl==3.1.2
vnstock_ezchart>=0.0.2
fake_useragent==1.5.1
cryptography==42.0.5
psutil==6.0.0
seaborn==0.13.2
fastapi==0.115.0
uvicorn[standard]==0.30.6